+++ src/components/ActivitiesList/ActivitiesList.js
@@ -1,26 +1,25 @@
-import Oops from '../Oops/Oops'
+import ErrorMessage from '../ErrorMessage/ErrorMessage'
 import FlipCard from '../flipCard/FlipCard'
 
 export default function ActivitiesList({ activities, onClickBookmark }) {
   if (!activities.length) {
-    return <Oops />
-  } else {
-    return activities.map(activity => (
-      <FlipCard
-        id={activity.id}
-        name={activity.name}
-        description={activity.description}
-        street={activity.street}
-        city={activity.city}
-        zipCode={activity.zipCode}
-        country={activity.country}
-        openingHours={activity.openingHours}
-        website={activity.website}
-        isFreeOfCharge={activity.isFreeOfCharge}
-        isBookmarked={activity.isBookmarked}
-        key={activity.id}
-        onClickBookmark={onClickBookmark}
-      />
-    ))
+    return <ErrorMessage />
   }
+  return activities.map(activity => (
+    <FlipCard
+      id={activity.id}
+      name={activity.name}
+      description={activity.description}
+      street={activity.street}
+      city={activity.city}
+      zipCode={activity.zipCode}
+      country={activity.country}
+      openingHours={activity.openingHours}
+      website={activity.website}
+      isFreeOfCharge={activity.isFreeOfCharge}
+      isBookmarked={activity.isBookmarked}
+      key={activity.id}
+      onClickBookmark={onClickBookmark}
+    />
+  ))
 }
